+++ src/controllers/ApiController.php
@@ -2818,7 +2818,7 @@ class ApiController extends Controller
             }
         }
 
-        return '0.6.2';
+        return '0.7.0';
     }
 
     private function getRequestId(): string