+++ src/services/ReadinessService.php
@@ -1613,7 +1613,7 @@ class ReadinessService extends Component
              }
         }
 
-        return '0.6.2';
+        return '0.7.0';
     }
 
     private function formatDate(?DateTimeInterface $date): ?string